+++ lib/linker.js
@@ -1,19 +1,9 @@
-const path = require("path");
-const unified = require("unified");
-const unifiedNgin = require("unified-engine");
 const uVisit = require("unist-util-visit");
-const remark_parse = require("remark-parse");
-const remark_slug = require("remark-slug");
 const remark_stringify = require("remark-stringify");
-const remark_link_headings = require("remark-autolink-headings");
-const remark_ref_links = require("remark-reference-links");
 const {link, html} = require("mdast-builder");
 
 const linkifyRegex = require("./ast/linkify");
-const {indexer} = require("./indexer");
-const {counter} = require("./counter");
-const {printAst, noopCompiler} = require("./ast/tools");
-const {getFileLinkUrl, toForwardSlash} = require("./path/tools");
+const {getFileLinkUrl} = require("./path/tools");
 /**
  * @typedef {import("./model/context")} Context
  * @typedef {import("./model/term")} Term
@@ -23,60 +13,8 @@ const api = {};
 
 // Tell remark_stringify to not produce any output for a term occurrence
 remark_stringify.Compiler.prototype.visitors["term-occurrence"] = () => {};
+remark_stringify.Compiler.prototype.visitors["image-occurrence"] = () => {};
 
-/**
- * Reads the pile of markdown files and replaces plaintext term occurrences with
- * links pointing to the term definition in one or more glossaries.
- *
- * @param {Context} context
- * @return {Promise<Context>} context
- */
-api.linkTermOccurrences = function(context) {
-    const {
-        baseDir, outDir, includeFiles, keepRawFiles,
-        excludeFiles, experimentalFootnotes, dev
-    } = context.opts;
-    return new Promise((resolve, reject) => {
-        const unifiedNginConf = {
-            processor: unified()
-                .use(remark_parse, { footnotes: experimentalFootnotes })
-                .use(printAst, { match: dev.printInputAst })  // Might be regex. /.*\/table\.md/g;
-                .use(remark_slug)
-                .use(linker, { context: context })
-                .use(remark_ref_links)
-                .use(remark_link_headings, {behavior: "wrap"})
-                .use(indexer, { context: context })
-                .use(counter, { context: context })
-                .use(printAst, { match: dev.printOutputAst })
-                .use(noopCompiler)
-                .use(remark_stringify)
-            ,cwd: baseDir
-            ,files: [
-                ...toForwardSlash(includeFiles)
-            ]
-            ,ignoreName: ".mdignore"
-            ,ignorePatterns: [
-                toForwardSlash(path.relative(baseDir, outDir))
-                ,...toForwardSlash(keepRawFiles)
-                ,...toForwardSlash(excludeFiles)
-            ]
-            ,extensions: ["md", "markdown", "mkd", "mkdn", "mkdown"]
-            ,alwaysStringify: true
-            ,output: false
-            ,out: false
-            ,color: true
-            ,silent: false
-        };
-        unifiedNgin(unifiedNginConf, (err, statusCode, uContext) => {
-            if (err) {
-                reject(err);
-            } else {
-                context.vFiles = [...context.vFiles, ...uContext.files];
-                resolve(context);
-            }
-        });
-    });
-};
 
 /**
  * Linker operates on the abstract syntax tree for a markdown file
@@ -85,13 +23,13 @@ api.linkTermOccurrences = function(context) {
  * @private
  * @param {{context: Context}} opts
  */
-function linker(opts) {
+api.linker = function(opts) {
     const {context} = opts;
     return (tree, vFile) => {
         uVisit(tree, getNodeVisitor(context, vFile));
         return tree;
     };
-}
+};
 
 /**
  * @private
