+++ lib/main.js
@@ -4,8 +4,7 @@ const GitHubSlugger = require("github-slugger");
 
 const Context = require("./model/context");
 const toForwardSlash = require("./path/tools").toForwardSlash;
-const glossarifier = require("./glossarifier");
-const linker = require("./linker");
+const reader = require("./reader");
 const writer = require("./writer");
 
 const CWD = proc.cwd();
@@ -14,8 +13,8 @@ function run(opts) {
     const context = new Context({opts: opts});
     prepare(context)
         .then(context => writer.copyBaseDirToOutDir(context))
-        .then(context => glossarifier.readGlossaries(context))
-        .then(context => linker.linkTermOccurrences(context))
+        .then(context => reader.readGlossaries(context))
+        .then(context => reader.readDocumentFiles(context))
         .then(context => writer.writeOutput(context))
         .then(context => writer.writeIndex(context))
         .then(context => writer.writeReport(context))
