+++ lib/terminator.js
@@ -1,91 +1,16 @@
 const path = require("path");
-const unified = require("unified");
-const unifiedNgin = require("unified-engine");
 const uVisit = require("unist-util-visit");
-const remark_parse = require("remark-parse");
-const remark_slug = require("remark-slug");
-const remark_link_headings = require("remark-autolink-headings");
 
-const Glossary = require("./model/glossary");
 const Term = require("./model/term");
-const {printAst, noopCompiler, getNodeText} = require("./ast/tools");
+const {getNodeText} = require("./ast/tools");
 const {toForwardSlash} = require("./path/tools");
 
 /**
  * @typedef {import('./model/Context')} Context
+ * @typedef {import('./model/glossary')} Glossary
  */
-
 
 const api = {};
-
-/**
- * Reads glossary markdown files into a dictionary.
- *
- * @param {Context} context
- * @return {Promise<Context>} context
- */
-api.readGlossaries = function(context) {
-    const {
-        baseDir, outDir, glossaries, keepRawFiles, excludeFiles,
-        experimentalFootnotes, dev
-    } = prepare(context).opts;
-
-    return new Promise((resolve, reject) => {
-        unifiedNgin(
-            {
-                processor: unified()
-                    .use(remark_parse, { footnotes: experimentalFootnotes })
-                    .use(printAst, { match: dev.printInputAst })  // Might be regex. /.*\/table\.md/g;
-                    .use(remark_slug)
-                    .use(remark_link_headings, {behavior: "wrap"})
-                    .use(terminator, { context: context })
-                    .use(printAst, { match: dev.printOutputAst }) // Might be regex. /.*\/table\.md/g;
-                    .use(noopCompiler)
-                ,cwd: baseDir
-                ,files: toForwardSlash(
-                    glossaries
-                        .map(g => g.file)
-                        .sort((f1, f2) => f1.localeCompare(f2, "en"))
-                )
-                ,ignoreName: ".mdignore"
-                ,ignorePatterns: [
-                    toForwardSlash(path.relative(baseDir, outDir))
-                    ,...toForwardSlash(keepRawFiles)
-                    ,...toForwardSlash(excludeFiles)
-                ]
-                ,extensions: ["md", "markdown", "mkd", "mkdn", "mkdown"]
-                ,alwaysStringify: true
-                ,output: false
-                ,out: false
-                ,color: true
-                ,silent: false
-            },
-            (err) => err ? reject(err) : resolve(context)
-        );
-    });
-};
-
-/**
- * @param {Context} context
- * @returns {Context}
- */
-function prepare(context) {
-    const {glossaries, baseDir, outDir} = context.opts;
-    // Instantiate `context.opts.glossaries` into an array of type `Glossary[]`.
-    // Thereby set up a `context.glossaries` map with keys being a glossary
-    // file path and value being a `Glossary` instance in order to access
-    // a `Glossary` instance by path. The latter allows to look up the glossary
-    // config by vFile metadata when proccesing glossary files with unifiedjs.
-    context.opts.glossaries = glossaries.map(conf => {
-        const g = new Glossary(conf);
-        g.file = conf.file || path.join(baseDir, "glossary.md");
-        g.basePath = toForwardSlash(path.resolve(baseDir, conf.file));
-        g.outPath  = toForwardSlash(path.resolve(outDir,  conf.file));
-        context.glossaries[g.basePath] = g;
-        return g;
-    });
-    return context;
-}
 
 /**
  * Terminator climbs along the shallow markdown syntax tree
@@ -97,7 +22,7 @@ function prepare(context) {
  * @param {{context: Context}} opts
  * @returns {(tree: Node, file: VFile) => Node} unifiedjs processor plug-in
  */
-function terminator(opts) {
+api.terminator = function(opts) {
     const { context } = opts;
     const { glossaries } = context;
     const { baseDir } = context.opts;
@@ -112,7 +37,7 @@ function terminator(opts) {
         uVisit(tree, getNodeVisitor(context, glossary));
         return tree;
     };
-}
+};
 
 
 /**
